Extract toggle_extension
